ue4-supervisely/docker.requirements.txt
Pillow==5.2.0
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.4
scikit-image==0.14.2
shapely>=1.6.4.post2
